Fix compiler warnings for ogrekit.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@7920 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Frameworks/OgreKit/Source/FindPanel/OgreHighlightThread.m
@@ -58,7 +58,7 @@
     OGRegularExpression *regex = [self regularExpression];
     
     /* blending highlight colors */
-    float   hue, saturation, brightness, alpha;
+    CGFloat   hue, saturation, brightness, alpha;
     [[[self highlightColor] colorUsingColorSpaceName:NSCalibratedRGBColorSpace] 
         getHue: &hue 
         saturation: &saturation
